+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.3.2
 
+### Features
+
+- Added support for managing [Aphlict][aphlict], Phabricator's real-time
+  notifications service.
+
 ### Bug Fixes
 
 - Added a warning if the installed PHP version is incompatible with Phabricator.
@@ -21,4 +26,5 @@ and starting again from scratch. The rewritten module has comprehensive unit
 tests (with [`rspec-puppet`](http://rspec-puppet.com)) and acceptance tests
 (with [`beaker-rspec`](https://github.com/puppetlabs/beaker-rspec)).
 
+[aphlict]: https://secure.phabricator.com/book/phabricator/article/notifications/
 [phd]: https://secure.phabricator.com/book/phabricator/article/managing_daemons/